+++ test/cli/output/test_default.py
@@ -87,7 +87,7 @@ def test_handle_initialized(capsys, execution_context, results_set, swagger_20):
     # And current directory
     assert f"rootdir: {os.getcwd()}" in lines
     # And number of collected operations
-    assert strip_style_win32(click.style("collected endpoints: 1", bold=True)) in lines
+    assert strip_style_win32(click.style("Collected API operations: 1", bold=True)) in lines
     # And the output has an empty line in the end
     assert out.endswith("\n\n")